feat(sdk/elixir): make dag instance be global
Instead of setting `dag` to the struct of the object type, use GenServer
to store it during invoke the function. And add import statement to
`Dagger.Mod.Object` to make a function can access `dag` instance via
`dag/0` function.

And this commit change how to calling a function by temporary passing
only arguments to the function (previously it pass a struct and
arguments).

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>